Add tests for recent cgo handling updates.
This commit adds tests for the recent cgo handling of char, unsigned char,
and uint8_t arrays.

In addition, it adds an architecture for the cgo specific testing based on
build constraints.  This was done because spew itself does not require
cgo, but in order to test its handling of the cgo types it needs to be fed
cgo data.  Rather than force all users to have a system which supports cgo
and an external compiler just to run the tests, the cgo test support must
explicitly be turned on via a build tag.